Enable S3 compatible storage
Adding endpoint, use_path_style_endpoint, and
ACL to the S3 settings (titled as "advanced").
This allows to use S3 compatible storage like Backblaze B2.

The AWS SDK for PHP sets the ACL to `private` as default,
which some providers don't support if the bucket policy
is set to something else. Therefore the `awsS3ACL` setting
allows to tweak that.