Fix fallback mime detection (#362)
The $this->mimeDetector returns an empty array when it does not match any file type. To actually fallback to finfo we have to check if the mime key is not present.

To reproduce, upload a simple txt file and observe an exception when the non existing mime key is accessed.